fix: add individual sorting algorithms for each type
