--- src/services/sort/types.ts
@@ -1,5 +1,0 @@
-export enum SortDir {
-	ASC = "asc",
-	DESC = "desc",
-	NONE = "default",
-}
